requirements.txt
torch~=2.7.1
pyyaml~=6.0.2